requirements.txt
selenium==3.141.0
pytest==6.2.2
webdriver-manager==1.7
allure-pytest==2.8
colorama==0.4.4
PyVirtualDisplay==2.0
pytest-html==3.1.0
mitmproxy==6.0.2

